requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pytube==15.0.0
pydub==0.25.1
speechrecognition==3.10.4
sqlalchemy==2.0.30
databases==0.9.0
psycopg2-binary
aiosqlite
googletrans==4.0.0-rc1
beautifulsoup4
python-decouple


